fix: GameCountU16 (#618)
* Fix: GameCountU16

item max count was limited to uint8, when you have uint16 on server.

* Update protocolgameparse.cpp

* Update item.h

+++ src/client/item.h
@@ -150,7 +150,7 @@ private:
     void internalDraw(int animationPhase, const Point& dest, const Color& color, bool drawThings, bool isMarked, LightView* lightView = nullptr);
     void setConductor();
 
-    uint8_t m_countOrSubType{ 0 };
+    uint16_t m_countOrSubType{ 0 };
 
     Color m_color{ Color::white };
